Give `StreamrWeb3` its own file, too
