+++ app/src/shared/flowtype/web3-types.js
@@ -1,7 +1,7 @@
 // @flow
 
 import type Transaction from '$shared/utils/Transaction'
-import { StreamrWeb3 } from '$shared/web3/web3Provider'
+import StreamrWeb3 from '$utils/web3/StreamrWeb3'
 import type TransactionError from '$shared/errors/TransactionError'
 import type { ProductId } from '$mp/flowtype/product-types'
